requirements.txt
google-genai==0.1.0
gradio==3.23.0